Coconut Tree added. Generates commonly in beaches and rarely in jungles. Blocks: Tree Crown (generates coconuts similar to Budding Amethyst); Log; Leaves;
currently missing plank blocks, stripped and all sided wood variants.

+

Items:
Coconut item (used as sapling and can be opened by right clicking with a sword or axe. Can also be thrown, where it will bounce once, and on the next hit it breaks and drops an Opened Coconut);
Opened Coconut (can be drank, restoring two drumsticks, leaving an Empty Coconut);
Empty Coconut (20% chance of filling a composter, can also be used as a fuel that melts two items)

Pink Sand Verbena added. A flower that generates commonly in beaches. It can be placed on sand.